R script for processing met data and convert to CLM-Microbe driving forces
Two R script files and two steps for data processing (gap filling and NetCDF file production); the algorithm for gap filling was simple in the current project, but it can be more complicated, depending on the data quality. This project was for CLM-Microbe simulation for the Howland AmeriFlux site (funded by a collaborative NSF project).